Integration tests refactoring

+++ SEPS/Acme.Seps.Domain.Subsidy/Entity/RenewableEnergySourceTariff.cs
@@ -14,8 +14,8 @@ namespace Acme.Seps.Domain.Subsidy.Entity
 
         protected RenewableEnergySourceTariff(
             ConsumerPriceIndex consumerPriceIndex,
-            int? lowerProductionLimit,
-            int? upperProductionLimit,
+            decimal? lowerProductionLimit,
+            decimal? upperProductionLimit,
             decimal lowerRate,
             decimal higherRate,
             Guid projectTypeId,
